Server-C43: Refactor project and add collation to find product by id feat.
