+++ src/Server/External/Infrastructure/Persistence/Relational/ODour.PostgresRelationalDb/Common/CommonConstant.cs
@@ -1,41 +1,45 @@
-namespace ODour.PostgresRelationalDb.Common
+namespace ODour.PostgresRelationalDb.Common;
+
+internal static class CommonConstant
 {
-    internal static class CommonConstant
+    internal static class DatabaseNativeType
     {
-        internal static class DatabaseNativeType
-        {
-            internal const string TEXT = "TEXT";
+        internal const string TEXT = "TEXT";
 
-            internal const string TIMESTAMPTZ = "TIMESTAMPTZ";
+        internal const string TIMESTAMPTZ = "TIMESTAMPTZ";
 
-            internal const string JSONB = "JSONB";
-        }
+        internal const string JSONB = "JSONB";
+    }
 
-        internal static class DatabaseSchemaName
-        {
-            internal const string MAIN = "main";
+    internal static class DatabaseSchemaName
+    {
+        internal const string MAIN = "main";
 
-            internal const string SLAVE = "slave";
+        internal const string SLAVE = "slave";
 
-            internal const string USER = "user";
+        internal const string USER = "user";
 
-            internal const string SYSTEM = "system";
+        internal const string SYSTEM = "system";
 
-            internal const string ROLE = "role";
+        internal const string ROLE = "role";
 
-            internal const string PRODUCT = "product";
+        internal const string PRODUCT = "product";
 
-            internal const string PAYMENT = "payment";
+        internal const string PAYMENT = "payment";
 
-            internal const string ORDER = "order";
+        internal const string ORDER = "order";
 
-            internal const string EVENT = "event";
+        internal const string EVENT = "event";
 
-            internal const string CATEGORY = "category";
+        internal const string CATEGORY = "category";
 
-            internal const string ACCOUNT_STATUS = "account_status";
+        internal const string ACCOUNT_STATUS = "account_status";
 
-            internal const string VOUCHER = "voucher";
-        }
+        internal const string VOUCHER = "voucher";
+    }
+
+    internal static class DatabaseCollation
+    {
+        public const string CASE_INSENSITIVE = "case_insensitive";
     }
 }
