+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/Auth/ConfirmUserEmail/Middlewares/ConfirmUserEmailValidationPreProcessor.cs
@@ -38,8 +38,6 @@ internal sealed class ConfirmUserEmailValidationPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
