+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/Auth/Login/Middlewares/LoginValidationPreProcessor.cs
@@ -34,8 +34,6 @@ internal sealed class LoginValidationPreProcessor : PreProcessor<LoginRequest, L
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
