+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/Auth/Logout/Middlewares/LogoutValidationPreProcessor.cs
@@ -34,8 +34,6 @@ internal sealed class LogoutValidationPreProcessor : PreProcessor<EmptyRequest, 
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
