+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/Auth/RefreshAccessToken/Middlewares/RefreshAccessTokenValidationPreProcessor.cs
@@ -38,8 +38,6 @@ internal sealed class RefreshAccessTokenValidationPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
