+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/Auth/RegisterAsUser/RegisterAsUserEndpoint.cs
@@ -4,7 +4,7 @@ using FastEndpoints;
 using Microsoft.AspNetCore.Http;
 using ODour.Application.Feature.Auth.RegisterAsUser;
 using ODour.FastEndpointApi.Feature.Auth.RegisterAsUser.HttpResponse;
-using ODour.FastEndpointApi.Feature.Auth.RegisterAsUser.Middlewares.Validation;
+using ODour.FastEndpointApi.Feature.Auth.RegisterAsUser.Middlewares;
 
 namespace ODour.FastEndpointApi.Feature.Auth.RegisterAsUser;
