+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/Auth/ResendUserConfirmationEmail/Middlewares/ResendUserConfirmationEmailValidationPreProcessor.cs
@@ -41,8 +41,6 @@ internal sealed class ResendUserConfirmationEmailValidationPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
