+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/Auth/ResetPassword/Middlewares/ResetPasswordValidationPreProcessor.cs
@@ -38,8 +38,6 @@ internal sealed class ResetPasswordValidationPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
