+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Cart/AddToCart/Middlewares/AddToCartValidationPreProcessor.cs
@@ -35,8 +35,6 @@ internal sealed class AddToCartValidationPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
