+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Cart/GetCartDetail/Middlewares/GetCartDetailCachingPreProcessor.cs
@@ -52,8 +52,6 @@ internal sealed class GetCartDetailCachingPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
