+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Cart/GetCartDetail/Middlewares/GetCartDetailValidationPreProcessor.cs
@@ -38,8 +38,6 @@ internal sealed class GetCartDetailValidationPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
