+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Cart/RemoveFromCart/Middlewares/RemoveFromCartValidationPreProcessor.cs
@@ -38,8 +38,6 @@ internal sealed class RemoveFromCartValidationPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
