+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Product/GetAllProducts/Middlewares/GetAllProductsCachingPreProcessor.cs
@@ -54,8 +54,6 @@ internal sealed class GetAllProductsCachingPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
