+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Product/GetAllProducts/Middlewares/GetAllProductsValidationPreProcessor.cs
@@ -38,8 +38,6 @@ internal sealed class GetAllProductsValidationPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
