+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Product/GetProductDetailByProductId/Middlewares/GetProductDetailByProductIdCachingPreProcessor.cs
@@ -60,8 +60,6 @@ internal sealed class GetProductDetailByProductIdCachingPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
