+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Product/GetProductDetailByProductId/Middlewares/GetProductDetailByProductIdValidationPreProcessor.cs
@@ -41,8 +41,6 @@ internal sealed class GetProductDetailByProductIdValidationPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
