+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Product/GetProductsForHomePage/Middlewares/GetProductsForHomePageCachingPreProcessor.cs
@@ -57,8 +57,6 @@ internal sealed class GetProductsForHomePageCachingPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
