+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Product/GetRelatedProductsByCategoryId/Middlewares/GetRelatedProductsByCategoryIdCachingPreProcessor.cs
@@ -62,8 +62,6 @@ internal sealed class GetRelatedProductsByCategoryIdCachingPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
