+++ src/Server/External/Presentation/ODour.FastEndpointApi/Feature/User/Product/GetRelatedProductsByCategoryId/Middlewares/GetRelatedProductsByCategoryIdValidationPreProcessor.cs
@@ -41,8 +41,6 @@ internal sealed class GetRelatedProductsByCategoryIdValidationPreProcessor
                 cancellation: ct
             );
 
-            context.HttpContext.MarkResponseStart();
-
             return;
         }
     }
